Add: 定期其他功能 (#35)
* 定期其他功能 通过测试

* 修复了lD错误的bug

* 删掉了TestInterrestService

Github我草泥马

* Revert "删掉了TestInterrestService"

This reverts commit dda05dbdcaf4f472cef3447a95d83b6d28d109b7.

* 同步运行测试

---------

+++ src/main/java/com/seg83/childbank/service/CurrentService.java
@@ -42,7 +42,7 @@ public class CurrentService {
      *
      * @param amount The amount to deposit.
      */
-    public void depositCurrentAccount(int amount) {
+    public void depositCurrentAccount(double amount) {
         double currentAmount = (Double) currentAccountDao.getAttribute("currentAccountAmount");
         double newAmount = currentAmount + amount;
         currentAccountDao.setAttribute("currentAccountAmount", newAmount);
@@ -98,4 +98,3 @@ public class CurrentService {
         return String.format("%.2f", rate) + "%";
     }
 }
-
